files : darmangeat.part7 & item

+++ Dos_Martires_TP2_java/Item4.java
@@ -1,5 +1,5 @@
 public class Item4 {
   public static void main(String[] args) {
-    System.out.println("resultat du calcul: " +(20.1 + 16.8));
+    System.out.println("resultat du calcul: " + ((float)(20.1 + 16.8)));
   }
 }
